requirements.txt
ipykernel
python-dotenv
google-api-python-client
polars
nltk
chromadb
sentence-transformers
fastapi[standard]